New features and fixes
Regex for random inputs in input_seed input module are available
(they work like a dictionary).

Return code of fuzzing module now is correct when a signal stops
the execution.

The pool object now is being closed every time is needed.

Crossover function has been replaced: now it takes a random point
and takes part of father and mother using the same point.

Mutator which generates chars now mutates with the same likelihood
that the other mutators (it was not correctly mutating).