Minor fixes and minor features
Instrumentation was not being correctly handled in the default
value (when no instrumentation was performed).

Now is being used the time it takes the executions for the reward
value of the genalg module.

Crossover function was not being correctly performed since it
could happen that the mother and father length was 0, what could
make to crash.

Tests fixed.